+++ Owl/Modules/Main/Chat/ChatMessage/ChatMessageView.swift
@@ -55,7 +55,7 @@ struct ChatMessageView: View {
             .padding(.horizontal, 8)
             .padding(.vertical, 4)
             .frame(width: screen.width, alignment: viewStore.alignment)
-            .onAppear { viewStore.send(.wasShown) }
+            .onAppear { viewStore.send(.onAppear) }
         }
         .background(
             Color(.systemGroupedBackground)
